refactor: pull out util function
